Added basic lua functions

+++ include/client.h
@@ -359,7 +359,7 @@ void NonNull resizeclient(Client *c, int16_t x, int16_t y, uint16_t width, uint1
  * 1 -> dont confide resize to monitor dimentions 
  * 0 -> confide resize within monitor dimentions
  */
-void NonNull resizemove(Client *c, int16_t x, int16_t y, uint8_t interact);
+void NonNull resizemove(Client *c, int32_t x, int32_t y, uint8_t interact);
 /* Sends a Protocol Event to specified client */
 void NonNull sendprotocolevent(Client *c, XCBAtom proto);
 /* Sets the flag "alwaysontop" to the provided Client */
